rename migration modules
