+++ protocols/migration/__init__.py
@@ -1,19 +1,19 @@
-from .migration import Migration2_1To3
+from .migration_reports_210_to_300 import Migration2_1To3
 
 from .base_migration import BaseMigration
 
-from .migration_participant_1_0_0_to_participant_1_0_3 import MigrationParticipants100To103
-from .migration_participant_1_0_3_to_participant_1_1_0 import MigrationParticipants103To110
-from .migration_reports_3_0_0_to_participant_1_0_0 import MigrationReports3ToParticipant1
-from .migration_participant_1_0_1_to_reports_3_0_0 import MigrationParticipants101ToReports
-from .migration_participant_1_0_3_to_participant_1_0_0 import MigrationParticipants103To100
-from .migration_participant_1_0_0_to_reports_3_0_0 import MigrationParticipants100ToReports
-from .migration_participant_1_1_0_to_participant_1_0_0 import MigrateParticipant110To100
-from .migration_participant_1_1_0_to_participant_1_0_3 import MigrateParticipant110To103
-from .migration_reports_4_0_0_to_reports_3_0_0 import MigrateReports400To300
+from .migration_participant_100_to_participant_103 import MigrationParticipants100To103
+from .migration_participant_103_to_participant_110 import MigrationParticipants103To110
+from .migration_reports_300_to_participant_100 import MigrationReports3ToParticipant1
+from .migration_participant_101_to_reports_300 import MigrationParticipants101ToReports
+from .migration_participant_103_to_participant_100 import MigrationParticipants103To100
+from .migration_participant_100_to_reports_300 import MigrationParticipants100ToReports
+from .migration_participant_110_to_participant_100 import MigrateParticipant110To100
+from .migration_participant_110_to_participant_103 import MigrateParticipant110To103
+from .migration_reports_400_to_reports_300 import MigrateReports400To300
 
-from .migration_reports_3_0_0_to_reports_4_0_0 import MigrateReports3To4
-from .migration_reports_4_0_0_to_reports_5_0_0 import MigrateReports400To500
-from .migration_reports_5_0_0_to_reports_4_0_0 import MigrateReports500To400
-from .migration_reports_5_0_0_to_reports_6_0_0 import MigrateReports500To600
+from .migration_reports_300_to_reports_400 import MigrateReports3To4
+from .migration_reports_400_to_reports_500 import MigrateReports400To500
+from .migration_reports_500_to_reports_400 import MigrateReports500To400
+from .migration_reports_500_to_reports_600 import MigrateReports500To600
 from .migration_helpers import MigrationHelpers
