+++ protocols/tests/test_migration/test_migration_reports_3_0_0_to_reports_4_0_0.py
@@ -6,7 +6,7 @@ from protocols.util.dependency_manager import VERSION_400
 from protocols.util.factories.avro_factory import GenericFactoryAvro
 from protocols.reports_3_0_0 import ReportEventCancer as ReportEventCancer_old
 from protocols.reports_4_0_0 import ReportEventCancer as ReportEventCancer_new
-from protocols.migration.migration_reports_3_0_0_to_reports_4_0_0 import MigrateReports3To4
+from protocols.migration.migration_reports_300_to_reports_400 import MigrateReports3To4
 import protocols.reports_3_0_0
